+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Open Graph type to `article` for blog post articles.
 - Open Graph site name.
 - Open Graph locale property.
+- Open Graph alternative locales.
 
 ### Changed