+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 ### Changed
 
 - To large summary card on Twitter.
+- Name of the site in the manifest file.
 
 ## [0.12.3] - 2021-03-21