+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Open Graph locale property.
 - Open Graph alternative locales.
 - Possibility to use Open Graph image that is set in Contentful.
+- Vector variant of the favicon that has different colour when device uses dark mode.
 
 ### Changed