+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Author information on author pages.
 - Author blog posts on author pages.
 - Page for all authors on the site.
+- Page for all categories on the site.
 
 ### Changed