+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
-## [Unreleased]
+## [0.13.0] - 2021-03-22
 
 ### Added
 
@@ -348,7 +348,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Support for internationalizing pages.
 - Both light and dark themes for the site.
 
-[unreleased]: https://github.com/visiosto/visiosto.fi/compare/v0.12.3...HEAD
+[unreleased]: https://github.com/visiosto/visiosto.fi/compare/v0.13.0...HEAD
+[0.13.0]: https://github.com/visiosto/visiosto.fi/compare/v0.12.3...v0.13.0
 [0.12.3]: https://github.com/visiosto/visiosto.fi/compare/v0.12.2...v0.12.3
 [0.12.2]: https://github.com/visiosto/visiosto.fi/compare/v0.12.1...v0.12.2
 [0.12.1]: https://github.com/visiosto/visiosto.fi/compare/v0.12.0...v0.12.1